use academies API client
