--- src/Frontend/Dfe.Complete/Services/ITrustService.cs
@@ -1,9 +1,0 @@
-﻿using Dfe.Complete.Application.Common.Models;
-
-namespace Dfe.Complete.Services
-{
-	public interface ITrustService
-	{
-        Task<IEnumerable<TrustDetailsDto>> GetTrustByUkprn(string ukprn);
-    }
-}
