+++ src/Frontend/Dfe.Complete/StartupConfiguration/DependencyConfigurationExtensions.cs
@@ -12,7 +12,6 @@ namespace Dfe.Complete.StartupConfiguration
             services.AddScoped<IErrorService, ErrorService>();
             services.AddScoped<CompleteApiClient, CompleteApiClient>();
             services.AddScoped<AcademiesApiClient, AcademiesApiClient>();
-            services.AddScoped<ITrustService, TrustService>();
 
             services.AddScoped<IAnalyticsConsentService, AnalyticsConsentService>();
