+++ src/Frontend/Dfe.Complete/StartupConfiguration/StartupConfigurationExtensions.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using Dfe.AcademiesApi.Client;
+using Dfe.AcademiesApi.Client.Contracts;
+using Dfe.TramsDataApi.Client.Extensions;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Dfe.Complete.StartupConfiguration
@@ -8,6 +11,7 @@ namespace Dfe.Complete.StartupConfiguration
         public static IServiceCollection AddCompleteClientProject(this IServiceCollection services, IConfiguration configuration)
         {
             services.AddClientDependencies();
+            services.AddAcademiesApiClient<ITrustsClient, TrustsClient>(configuration);
 
             return services;
         }
